Remove CSCCommonTrigger
