--- DQM/L1TMonitor/python/L1TCSCTF_cff.py
@@ -1,5 +1,0 @@
-import FWCore.ParameterSet.Config as cms
-
-from DQM.L1TMonitor.L1TCSCTF_cfi import *
-from L1Trigger.CSCCommonTrigger.CSCCommonTrigger_cfi import *
-
