--- L1Trigger/CSCCommonTrigger/BuildFile.xml
@@ -1,5 +1,0 @@
-<export>
-  <lib name="1"/>
-</export>
-<use name="boost"/>
-<use name="clhep"/>
