+++ L1Trigger/CSCTrackFinder/python/csctfDigis_cfi.py
@@ -1,6 +1,6 @@
 import FWCore.ParameterSet.Config as cms
 
-from L1Trigger.CSCCommonTrigger.CSCCommonTrigger_cfi import *
+from L1Trigger.CSCTriggerPrimitives.CSCCommonTrigger_cfi import *
 csctfDigis = cms.EDProducer("CSCTFCandidateProducer",
     CSCTrackProducer = cms.untracked.InputTag("csctfTrackDigis"),
     MuonSorter = cms.PSet(
