+++ L1Trigger/CSCTrackFinder/src/CSCSectorReceiverMiniLUT.cc
@@ -1,5 +1,4 @@
-#include <L1Trigger/CSCCommonTrigger/interface/CSCPatternLUT.h>
-#include <L1Trigger/CSCCommonTrigger/interface/CSCFrontRearLUT.h>
+#include <L1Trigger/CSCTriggerPrimitives/interface/CSCPatternBank.h>
 #include <DataFormats/L1CSCTrackFinder/interface/CSCBitWidths.h>
 #include <DataFormats/L1CSCTrackFinder/interface/CSCTFConstants.h>
 #include <DataFormats/L1TMuon/interface/CSCConstants.h>
