+++ L1Trigger/CSCTriggerPrimitives/BuildFile.xml
@@ -8,7 +8,7 @@
 <use name="FWCore/Utilities"/>
 <use name="Geometry/CSCGeometry"/>
 <use name="Geometry/GEMGeometry"/>
-<use name="L1Trigger/CSCCommonTrigger"/>
+<use name="Geometry/CSCGeometry"/>
 <export>
   <lib name="1"/>
 </export>
