+++ L1Trigger/CSCTriggerPrimitives/plugins/BuildFile.xml
@@ -7,7 +7,6 @@
 <use name="FWCore/ParameterSet"/>
 <use name="FWCore/Utilities"/>
 <use name="Geometry/CSCGeometry"/>
-<use name="L1Trigger/CSCCommonTrigger"/>
 <use name="L1Trigger/CSCTriggerPrimitives"/>
 <use name="SimDataFormats/TrackingHit"/>
 <use name="SimDataFormats/GeneratorProducts"/>
