+++ L1Trigger/L1TMuon/BuildFile.xml
@@ -9,7 +9,7 @@
 <use name="DataFormats/RPCRecHit"/>
 <use name="FWCore/PluginManager"/>
 <use name="L1Trigger/L1TCommon"/>
-<use name="L1Trigger/CSCCommonTrigger"/>
+<use name="L1Trigger/CSCTriggerPrimitives"/>
 <use name="L1Trigger/DTUtilities"/>
 <use name="Geometry/Records"/>
 <use name="Geometry/CSCGeometry"/>
