+++ L1Trigger/L1TMuon/src/GeometryTranslator.cc
@@ -13,7 +13,7 @@
 
 #include "L1Trigger/DTUtilities/interface/DTTrigGeom.h"
 #include "DataFormats/L1TMuon/interface/CSCConstants.h"
-#include "L1Trigger/CSCCommonTrigger/interface/CSCPatternLUT.h"
+#include "L1Trigger/CSCTriggerPrimitives/interface/CSCPatternBank.h"
 
 #include "L1Trigger/L1TMuon/interface/MuonTriggerPrimitive.h"
 
@@ -247,7 +247,7 @@ GlobalPoint GeometryTranslator::getCSCSpecificPoint(const TriggerPrimitive& tp) 
   double offset = 0.0;
   switch (1) {
     case 1:
-      offset = CSCPatternLUT::get2007Position(pattern);
+      offset = CSCPatternBank::getLegacyPosition(pattern);
   }
   const unsigned halfstrip_offs = static_cast<unsigned>(0.5 + halfstrip + offset);
   const unsigned strip = halfstrip_offs / 2 + 1;  // geom starts from 1
