+++ Utilities/ReleaseScripts/scripts/git-publish
@@ -555,7 +555,7 @@ CMSSW_CATEGORIES={"Operations": ["Configuration/Applications",
   "L1TriggerConfig/L1GeometryProducers", "L1TriggerConfig/L1GtConfigProducers",
   "L1TriggerConfig/L1ScalesProducers", "L1TriggerConfig/RCTConfigProducers",
   "L1TriggerConfig/RPCTriggerConfig", "L1TriggerOffline/Configuration",
-  "L1TriggerOffline/L1Analyzer", "L1Trigger/CSCCommonTrigger",
+  "L1TriggerOffline/L1Analyzer",
   "L1Trigger/CSCTrackFinder", "L1Trigger/CSCTriggerPrimitives",
   "L1Trigger/Configuration", "L1Trigger/DTBti", "L1Trigger/DTSectorCollector",
   "L1Trigger/DTTrackFinder", "L1Trigger/DTTraco", "L1Trigger/DTTrigger",
